requirements.txt
streamlit>=1.32.0
python-dotenv>=1.0.0
langchain-openai>=0.0.2
mcp-use>=0.1.0
pytz>=2024.1
